refactor: create services folder, move commands out of server file
