+++ packages/server/lib/commands/destroyView.js
@@ -1,0 +1,5 @@
+exports.default = {
+  handler: async ({sid, vId, setView}) => {
+    setView(sid, vId, null);
+  }
+};
