+++ packages/server/lib/commands/getUsers.js
@@ -1,0 +1,7 @@
+exports.default = {
+  handler: async ({priv, sid, config, sendUsers}) => {
+    if (priv && !config.public) {
+      sendUsers(sid);
+    }
+  }
+};
