+++ packages/server/lib/index.js
@@ -1,13 +1,13 @@
-exports.cfg = require("./cfg");
+exports.cfg = require("./services/cfg");
 
-exports.cookies = require("./cookies");
-exports.csrf = require("./csrf");
-exports.db = require("./db");
-exports.filetree = require("./filetree");
-exports.log = require("./log");
-exports.manifest = require("./manifest");
-exports.paths = require("./paths");
-exports.resources = require("./resources");
-exports.server = require("./server");
-exports.svg = require("./svg");
-exports.utils = require("./utils");
+exports.cookies = require("./services/cookies");
+exports.csrf = require("./services/csrf");
+exports.db = require("./services/db");
+exports.filetree = require("./services/filetree");
+exports.log = require("./services/log");
+exports.manifest = require("./services/manifest");
+exports.paths = require("./services/paths");
+exports.resources = require("./services/resources");
+exports.server = require("./services/server");
+exports.svg = require("./services/svg");
+exports.utils = require("./services/utils");
