+++ packages/server/lib/services/filetree.js
