+++ packages/server/lib/services/svg.js
